refactor: categorize maskrcnn files into different folders
